refactor: Move API schemas to dedicated schemas directory
- Create backend/app/schemas/file.py for file-related request/response schemas
- Move FileUploadRequest, BulkUploadRequest, FileConfirmation, BulkConfirmUploadRequest
- Update imports in files.py route handler
- Follows FastAPI convention for separating API contracts from routes
- Improves code organization and reusability
