+++ backend/app/schemas/__init__.py
@@ -1,0 +1,13 @@
+from .file import (
+    BulkConfirmUploadRequest,
+    BulkUploadRequest,
+    FileConfirmation,
+    FileUploadRequest,
+)
+
+__all__ = [
+    "FileUploadRequest",
+    "BulkUploadRequest",
+    "FileConfirmation",
+    "BulkConfirmUploadRequest",
+]
